+++ blocks/form/src/components/BooleanInput/index.tsx
@@ -4,6 +4,8 @@ import classNames from 'classnames';
 import { VNode } from 'preact';
 
 import { BooleanField, InputProps } from '../../../block.js';
+import { getValueByNameSequence } from '../../utils/getNested.js';
+import { isRequired } from '../../utils/requirements.js';
 
 type BooleanInputProps = InputProps<boolean, BooleanField>;
 
@@ -16,16 +18,17 @@ export function BooleanInput({
   disabled,
   error,
   field,
+  formValues,
   name,
   onChange,
   readOnly,
-  required,
-  value,
 }: BooleanInputProps): VNode {
   const { utils } = useBlock();
   const { color, icon, inline, label, labelText, size, switch: switchType, tag } = field;
 
+  const value = getValueByNameSequence(name, formValues);
   const checkboxLabel = utils.remap(label, value);
+  const required = isRequired(field);
 
   return (
     <CheckboxField